refactor: extract security headers middleware to separate module
